chore: Add Docker and Gradle workflows
Introduced two new GitHub Actions workflows. The first workflow is for building and pushing Docker images, which triggers on push events to the main branch or when a new tag is created. It includes steps for setting up JDK, Gradle, building with Gradle Wrapper, uploading WAR artifact, setting up QEMU and Docker Buildx, logging into GitHub Container Registry, extracting metadata, and finally building and pushing the image.

The second workflow is for Java CI with Gradle. This workflow triggers on both push events to the main branch as well as pull requests targeting the main branch. It includes steps for setting up JDK and Gradle, building with Gradle Wrapper, and generating & submitting a dependency graph.